+++ othoz/paragraph/tests/test_session.py
@@ -21,7 +21,8 @@ def graph():
 
 
 class TestForwardGenerator:
-    def test_generated_values(self, graph):
+    @staticmethod
+    def test_generated_values(graph):
         items = list(traverse_fw(graph.output))
         expected = [graph.input] + graph.output
 
@@ -29,7 +30,8 @@ class TestForwardGenerator:
 
 
 class TestBackwardGenerator:
-    def test_generated_value_wo_boundary(self, graph):
+    @staticmethod
+    def test_generated_value_wo_boundary(graph):
         items = list(traverse_bw(graph.output))
         expected = [graph.output[1], graph.output[0], graph.input]
 
@@ -39,7 +41,8 @@ class TestBackwardGenerator:
 @pytest.mark.parametrize("max_workers", [pytest.param(1, id="Single thread"),
                                          pytest.param(50, id="Multi-threaded")])
 class TestEvaluate:
-    def test_variable_evaluation_is_correct(self, graph, max_workers):
+    @staticmethod
+    def test_variable_evaluation_is_correct(graph, max_workers):
         res = evaluate(graph.output, args={graph.input: "Input value"}, max_workers=max_workers)
 
         assert res[0] == "op0 return value"
@@ -48,7 +51,8 @@ class TestEvaluate:
         graph.output[0].op._run.assert_called_once_with(arg="Input value")
         graph.output[1].op._run.assert_called_once_with(arg0="Input value", arg1="op0 return value")
 
-    def test_evaluation_is_lazy(self, graph, max_workers):
+    @staticmethod
+    def test_evaluation_is_lazy(graph, max_workers):
         res = evaluate([graph.output[0]], args={graph.input: "Input value"}, max_workers=max_workers)
         operation = graph.output[1].op
 
@@ -59,18 +63,21 @@ class TestEvaluate:
 @pytest.mark.parametrize("max_workers", [pytest.param(1, id="Single thread"),
                                          pytest.param(50, id="Multi-threaded")])
 class TestApply:
-    def test_apply_returns_correct_number_of_results(self, graph, max_workers):
+    @staticmethod
+    def test_apply_returns_correct_number_of_results(graph, max_workers):
         res = list(apply(graph.output, args={}, iter_args=[{graph.input: "Input value"}] * 5, max_workers=max_workers))
 
         assert len(res) == 5
 
-    def test_apply_raises_on_overwriting_an_input_variable(self, graph, max_workers):
+    @staticmethod
+    def test_apply_raises_on_overwriting_an_input_variable(graph, max_workers):
         with pytest.raises(ValueError):
             list(apply(graph.output, args={graph.input: "Input value"}, iter_args=[{graph.input: "Input value"}] * 5, max_workers=max_workers))
 
 
 class TestRequirementSolving:
-    def test_req_update_func_called(self):
+    @staticmethod
+    def test_req_update_func_called():
         operation = mock_op()
 
         var = Variable()
@@ -82,4 +89,4 @@ class TestRequirementSolving:
         expected = {res: output_req, var: MockReq("Arg requirement")}
 
         assert reqs == expected
-        res.op._arg_requirements.assert_called_once_with(output_req, "b")
+        res.op.arg_requirements.assert_called_once_with(output_req, "b")